[Feature] Support Business Logic level dependency conditions
This pull request makes it possible to mix business logic with Surveyor, not being tied to the survey itself.

To achieve this, we've extended the `Surveyor::DependencyCondition` model with two extra fields:

- `condition_type` - "default" by default, but can be passed as logic to match any application-level logic
- `logic_reference` - The name of the logic variable to be associated with the dependency condition

Here's an example on how it can be used on the DSL:

```
  q_0 "Do you think you'll be saved?"
  a_0 "Yes"
  a_1 "No"
  dependency :rule => "A"
  condition_A "religious", "==", { condition_type: 'logic', text_value: "true" }
```

Where on the first parameter of `condition_A` we specify the `logic_reference`, on the second the operator (as usual) and then an hash stating the condition type and the value to be matched with.

On this example, the application would handle the value of "religious" on their dependency management code, only presenting this question to a user that fits in.