[Feature] Properly parse Validation Conditions
Validation Conditions were created not to only validate the format of a specific response, but also to validate it based on a question/answer combination.

Currently the parser was not mapping out the question/answer references, so we just replicated what was being done with the DependencyConditons.

So now we'll be able to create specific validations like the following:

```
      q_3 'Number of other people in the vehicle:',
          display_type: 'number_input',
          common_identifier: 'car_travel_party_size',
          is_mandatory: true
      a_0 '&nbsp;'

      q_4 'How many household members were with you?',
          display_type: 'number_input',
          common_identifier: 'hh_no_members',
          is_mandatory: true
      dependency rule: 'A'
      condition_A :q_3, '>', { text_value: "0", answer_reference: "0" }
      a_0 '&nbsp;'
      validation rule: 'A'
      condition_A "<=", { question_reference: "3", answer_reference: "0" }
```

Where the second question will include a validation condition, where its value should be bigger or equal to the first question response value.